+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.2
 ### 2021/07/16
 - edit code and docs style
-
+- edit multiSelector mode to picker component
 ### 2021/07/15
 - add selector mode to picker component